+++ test/plugin/determine_mime_type_test.rb
@@ -193,6 +193,14 @@ describe Shrine::Plugins::DetermineMimeType do
       assert_equal "foo/bar", @shrine.determine_mime_type(fakeio(content_type: "foo/bar"))
     end
 
+    it "strips away any media type parameters" do
+      assert_equal "foo/bar", @shrine.determine_mime_type(fakeio(content_type: "foo/bar; param=value"))
+    end
+
+    it "returns nil when #content_type is nil" do
+      assert_nil @shrine.determine_mime_type(fakeio(content_type: nil))
+    end
+
     it "returns nil when IO doesn't respond to #content_type" do
       assert_nil @shrine.determine_mime_type(image)
     end